% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champy_densidad.R
\name{champy_densidad}
\alias{champy_densidad}
\title{Probability Density Function of the Champy Distribution}
\usage{
champy_densidad(x)
}
\arguments{
\item{x}{A numeric value or vector where the density function is evaluated.}
}
\value{
A numeric vector with the corresponding density values.
}
\description{
Evaluates the probability density function (PDF) of the Champy
continuous distribution at given values of \code{x}.
}
\details{
The Champy distribution is defined on the interval [-1, 1] with
probability density function:

\deqn{
f(x) =
\begin{cases}
\frac{15}{16}(1 - x^2)^2, & -1 \leq x \leq 1 \\
0, & \text{otherwise}
\end{cases}
}

This function is vectorized and returns 0 for values outside the support.
}
\examples{
champy_densidad(0)
champy_densidad(c(-1, 15/12, 1))

}
