% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/champy_random.R
\name{champy_random}
\alias{champy_random}
\title{Random Sample Generator for the Champy Distribution}
\usage{
champy_random(n)
}
\arguments{
\item{n}{An integer specifying the number of random samples to generate.}
}
\value{
A numeric vector of length \code{n} containing random samples
drawn from the Champy distribution.
}
\description{
Generates random samples from the Champy continuous distribution
using the acceptance-rejection method with a constant majorant.
}
\details{
The Champy distribution has probability density function:

\deqn{
f(x) =
\begin{cases}
\frac{15}{16}(1 - x^2)^2, & -1 \leq x \leq 1 \\
0, & \text{otherwise}
\end{cases}
}

This function implements the acceptance-rejection algorithm:
\enumerate{
  \item Generate a candidate \code{x} from a uniform distribution on [-1, 1].
  \item Generate a uniform random variable \code{u} in [0, 1].
  \item Accept \code{x} if \code{u <= f(x)/m(x)}, where \code{m(x) = 15/16}.
  \item Repeat until \code{n} samples are accepted.
}
}
